move alertAvailableFor to an object for alerting
